Automate arm toolchain things

+++ build.zig
@@ -10,6 +10,40 @@ pub fn build(b: *std.Build) void {
 
     const target = b.resolveTargetQuery(query);
 
+    // Teensy compile
+    const compile_dep = if (builtin.os.tag == .windows)
+        b.dependency("teensy_compile_windows", .{})
+    else
+        @panic("Unsupported OS");
+    var compile_install_dir = if (builtin.os.tag == .windows)
+        b.addInstallDirectory(.{
+            .source_dir = compile_dep.path(""),
+            .install_dir = .bin,
+            .install_subdir = "compile",
+        })
+    else
+        b.addInstallDirectory(.{
+            .source_dir = compile_dep.path(""),
+            .install_dir = .bin,
+            .install_subdir = "compile",
+        });
+
+    const compile_dir = b.getInstallPath(
+        compile_install_dir.options.install_dir,
+        compile_install_dir.options.install_subdir,
+    );
+    const arm_dir = b.pathJoin(&[_][]const u8{ compile_dir, "arm" });
+    // const arm_cc_path = b.pathJoin(
+    //     &[_][]const u8{ arm_dir, if (builtin.os.tag == .windows) "arm-none-eabi-gcc.exe" else "arm-none-eabi-gcc" },
+    // );
+    // const arm_cpp_path = b.pathJoin(
+    //     &[_][]const u8{ arm_dir, if (builtin.os.tag == .windows) "arm-none-eabi-g++.exe" else "arm-none-eabi-g++" },
+    // );
+    const arm_objcopy_path = b.pathJoin(
+        &[_][]const u8{ arm_dir, "bin", if (builtin.os.tag == .windows) "arm-none-eabi-objcopy.exe" else "arm-none-eabi-objcopy" },
+    );
+
+    // Main executable — the actual firmware implementation
     const exe_mod = b.createModule(.{
         .root_source_file = b.path("src/_startup.zig"),
         .target = target,
@@ -18,6 +52,7 @@ pub fn build(b: *std.Build) void {
     });
 
     exe_mod.strip = false;
+    exe_mod.addSystemIncludePath(.{ .cwd_relative = b.pathJoin(&[_][]const u8{ arm_dir, "arm-none-eabi", "include" }) });
 
     const exe = b.addExecutable(.{
         .name = "teensy_zig",
@@ -26,12 +61,27 @@ pub fn build(b: *std.Build) void {
 
     exe.setLinkerScript(b.path("linker_script.ld"));
     exe.entry = .{ .symbol_name = "__ivt_start" };
+    exe.step.dependOn(&compile_install_dir.step);
 
     b.installArtifact(exe);
 
+    // Compile to IHEX format
+    const objcopy_run = b.addSystemCommand(&[_][]const u8{arm_objcopy_path});
+    objcopy_run.addArgs(&[_][]const u8{ "--output-target", "ihex", "--remove-section", ".eeprom" });
+    objcopy_run.addFileArg(exe.getEmittedBin());
+    _ = objcopy_run.addArg(b.getInstallPath(.bin, "teensy_zig.hex"));
+
+    var hex_step = b.step("hex", "Compile firmware to hex file");
+    hex_step.dependOn(b.getInstallStep());
+    hex_step.dependOn(&compile_install_dir.step);
+    hex_step.dependOn(&objcopy_run.step);
+
     // Teensy tools
     var tools_step = b.step("tools", "Install teensy tools");
-    const tools_dep = b.dependency("teensy_tools", .{});
+    const tools_dep = if (builtin.os.tag == .windows)
+        b.dependency("teensy_tools_windows", .{})
+    else
+        @panic("Unsupported OS");
     var tools_install_dir = if (builtin.os.tag == .windows)
         b.addInstallDirectory(.{
             .source_dir = tools_dep.path(""),
@@ -75,22 +125,24 @@ pub fn build(b: *std.Build) void {
             if (builtin.os.tag == .windows) "teensy_post_compile.exe" else "teensy_post_compile",
         },
     );
-    const upload_run = b.addSystemCommand(&[_][]const u8{teensy_post_compile_path});
 
-    // FIXME: the teensy_zig.hex should be in the install directory, but currently it is build manually and located
-    // at the root
-    upload_run.addPrefixedDirectoryArg("-path=", b.path(""));
-    // FIXME: this should not be hardcoded
+    const upload_run = b.addSystemCommand(&[_][]const u8{teensy_post_compile_path});
+    const upload_reboot_option = b.option(bool, "upload-reboot", "Reboot Teensy after upload. Default to true") orelse true;
+    if (upload_reboot_option) {
+        upload_run.addArg("-reboot");
+    }
+    upload_run.addArg(b.fmt("-path={s}", .{b.getInstallPath(.bin, "")}));
     upload_run.addArg("-file=teensy_zig");
     upload_run.addArg(b.fmt("-tools={s}", .{tools_dir}));
-    // FIXME: should be an option
-    upload_run.addArg("-reboot");
     upload_run.addArg("-v");
 
     const upload_step = b.step("upload", "Upload firmware to Teensy board");
     upload_step.dependOn(&upload_run.step);
-    const upload_port_opton = b.option([]const u8, "upload-port", "USB port to upload to");
-    if (upload_port_opton) |port| {
+    upload_step.dependOn(tools_step);
+    upload_step.dependOn(hex_step);
+
+    const upload_port_option = b.option([]const u8, "upload-port", "USB port to upload to");
+    if (upload_port_option) |port| {
         upload_run.addArg(b.fmt("-port={s}", .{port}));
     }
